FAI-243 - provided Type dependent random value generation
